Refactor extract CategoryBadgesContainer to component
