+++ src/components/user/product/ProductCard.tsx
@@ -2,7 +2,7 @@ import { Product } from '@@types/index';
 import styled from '@emotion/styled';
 import AppLabel from '@components/common/label/AppLabel';
 import React, { useState } from 'react';
-import ProductDialog from '@components/product/ProductDialog';
+import ProductDialog from '@components/user/product/ProductDialog';
 
 interface ProductCardProps {
   product: Product;
